feat(product): 使用BigDecimal优化价格精度
- 将所有产品相关DTO中的价格字段由Double改为BigDecimal
- 为ProductStockUpdateDTO添加无参构造函数支持
- 引入java.math.BigDecimal包以支持高精度计算
- 确保价格数据在传输过程中保持精度完整性- 更新价格验证逻辑适配新的数据类型
- 统一产品模块中涉及价格的所有数据结构定义

+++ src/main/java/com/community_shop/backend/dto/product/ProductQueryDTO.java
@@ -7,6 +7,8 @@ import com.community_shop.backend.enums.SortEnum.SortDirectionEnum;
 import com.community_shop.backend.enums.CodeEnum.ProductConditionEnum;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * 商品搜索条件DTO（匹配ProductService.selectProductByKeyword及Mapper层selectByCondition）
  */
@@ -20,10 +22,10 @@ public class ProductQueryDTO extends PageParam {
     private String category;
 
     /** 最低价格（价格区间下限） */
-    private Double minPrice;
+    private BigDecimal minPrice;
 
     /** 最高价格（价格区间上限） */
-    private Double maxPrice;
+    private BigDecimal maxPrice;
 
     /** 商品成色（精确筛选，如“全新”“9成新”） */
     private ProductConditionEnum condition;
